+++ kernels/linux-5.15.204/linux-5.15.204/net/ipv4/inet_connection_sock.c
@@ -117,7 +117,7 @@ bool inet_rcv_saddr_any(const struct sock *sk)
 	return !sk->sk_rcv_saddr;
 }
 
-void inet_get_local_port_range(struct net *net, int *low, int *high)
+void inet_get_local_port_range(struct net *net, u64 *low, u64 *high)
 {
 	unsigned int seq;
